+++ src/assets/svg/icons/DownloadIcon.vue
@@ -1,0 +1,9 @@
+/* Font Awesome Pro 6.3.0 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license
+(Commercial License) Copyright 2023 Fonticons, Inc. */
+<template>
+	<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 448 512" fill="currentColor" width="1em" height="1em">
+		<path
+			d="M64 480V416H0v64H64zm96 0V416H96v64h64zm32 0h64V416H192v64zm160 0V416H288v64h64zm32 0h64V416H384v64zM256 160v48h48 28.1L224 316.1 115.9 208H144h48V160 80h64v80zm1.9 190.1L384 224V208 160H352 336 304V112 80 32H256 192 144V80v32 48H112 96 64v48 16L190.1 350.1 224 384l33.9-33.9z"
+		/>
+	</svg>
+</template>