+++ worker.vite.config.ts
@@ -1,10 +1,17 @@
+import { displayName as name, version } from "./package.json";
 import path from "path";
 import { defineConfig, loadEnv } from "vite";
 
 const r = (...args: string[]) => path.resolve(__dirname, ...args);
 
 export default defineConfig(({ mode }) => {
-	process.env = { ...process.env, ...loadEnv(mode, process.cwd()) };
+	process.env = {
+		...process.env,
+		...loadEnv(mode, process.cwd()),
+		VITE_APP_NAME: name,
+		VITE_APP_VERSION: version,
+		VITE_APP_VERSION_BRANCH: process.env.BRANCH,
+	};
 
 	process.stdout.write("Building worker...\n");